Change Ids from type Integer to Long

+++ src/main/java/seng468/scalability/matchingEngine/MatchingEngineOrdersRepository.java
@@ -18,32 +18,32 @@ public interface MatchingEngineOrdersRepository extends JpaRepository<StockOrder
     @Modifying
     @Transactional
     @Query("DELETE FROM StockOrder so WHERE so.stock_tx_id = ?1")
-    void deleteByStock_tx_id(Integer stock_tx_id);
+    void deleteByStock_tx_id(Long stock_tx_id);
 
 
     @Query("SELECT so FROM StockOrder so WHERE so.stock_id = ?1 AND so.orderStatus != seng468.scalability.models.entity.StockOrder$OrderStatus.COMPLETED" +
             " AND so.expired = false AND so.is_buy = false ORDER BY so.price ASC NULLS FIRST, so.timestamp ASC")
-    LinkedList<StockOrder> getAllSellByStock_id(int stock_id);
+    LinkedList<StockOrder> getAllSellByStock_id(Long stock_id);
 
     @Query("SELECT so FROM StockOrder so WHERE so.stock_id = ?1 AND so.orderStatus != seng468.scalability.models.entity.StockOrder$OrderStatus.COMPLETED " +
             "AND so.expired = false AND so.is_buy = true ORDER BY so.price DESC NULLS FIRST, so.timestamp ASC")
-    LinkedList<StockOrder> getAllBuyByStock_id(int stock_id);
+    LinkedList<StockOrder> getAllBuyByStock_id(Long stock_id);
 
     @Query("SELECT so from StockOrder so WHERE so.username = ?1")
     List<StockOrder> findAllByUsername(String username);
 
 
     @Query("SELECT so from StockOrder so where so.stock_tx_id = ?1 AND so.username = ?2")
-    StockOrder findByUsernameAndStockTxId(Integer stockTxId, String username);
+    StockOrder findByUsernameAndStockTxId(Long stockTxId, String username);
 
     //rework this!
     @Query("SELECT so FROM StockOrder so WHERE so.stock_id = ?1 AND so.orderStatus = seng468.scalability.models.entity.StockOrder$OrderStatus.COMPLETED" +
             " AND so.price is not null AND so.expired = false AND so.is_buy = false ORDER BY so.timestamp DESC")
-    LinkedList<StockOrder> getLastSellOrderByStockId(int stock_id, Pageable pageable);
+    LinkedList<StockOrder> getLastSellOrderByStockId(Long stock_id, Pageable pageable);
 
     @Query("SELECT so FROM StockOrder so WHERE so.stock_id = ?1 AND so.orderStatus != seng468.scalability.models.entity.StockOrder$OrderStatus.COMPLETED" +
             " AND so.price is not null AND so.expired = false AND so.is_buy = false ORDER BY so.price DESC")
-    LinkedList<StockOrder> getLowestSellOrderByStockId(int stock_id, Pageable pageable);
+    LinkedList<StockOrder> getLowestSellOrderByStockId(Long stock_id, Pageable pageable);
 
 
     @Query("SELECT so from StockOrder so WHERE so.orderType = seng468.scalability.models.entity.StockOrder$OrderType.LIMIT AND " +
@@ -51,5 +51,5 @@ public interface MatchingEngineOrdersRepository extends JpaRepository<StockOrder
     LinkedList<StockOrder> getAllLimitOrders();
 
     @Query("SELECT so from StockOrder so WHERE so.parent_stock_tx_id = ?1")
-    LinkedList<StockOrder> findAllParentChildTransactions(Integer parentStockTxId);
+    LinkedList<StockOrder> findAllParentChildTransactions(Long parentStockTxId);
 }
